# backend/internal/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Arrflix API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/name-templates": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "name-templates"
                ],
                "summary": "List name templates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.nameTemplateSwagger"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "name-templates"
                ],
                "summary": "Create name template",
                "parameters": [
                    {
                        "description": "Create name template",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.NameTemplateCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.nameTemplateSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/name-templates/default/{type}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "name-templates"
                ],
                "summary": "Get default name template by type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template type (movie or series)",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.nameTemplateSwagger"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/name-templates/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "name-templates"
                ],
                "summary": "Get name template",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.nameTemplateSwagger"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "name-templates"
                ],
                "summary": "Update name template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update name template",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.NameTemplateUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.nameTemplateSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "name-templates"
                ],
                "summary": "Delete name template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.NameTemplateCreateRequest": {
            "type": "object",
            "required": [
                "default",
                "movie_dir_template",
                "name",
                "series_season_template",
                "series_show_template",
                "template",
                "type"
            ],
            "properties": {
                "default": {
                    "type": "boolean"
                },
                "movie_dir_template": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "series_season_template": {
                    "type": "string"
                },
                "series_show_template": {
                    "type": "string"
                },
                "template": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "handlers.NameTemplateUpdateRequest": {
            "type": "object",
            "required": [
                "default",
                "movie_dir_template",
                "name",
                "series_season_template",
                "series_show_template",
                "template",
                "type"
            ],
            "properties": {
                "default": {
                    "type": "boolean"
                },
                "movie_dir_template": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "series_season_template": {
                    "type": "string"
                },
                "series_show_template": {
                    "type": "string"
                },
                "template": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "handlers.nameTemplateSwagger": {
            "type": "object",
            "required": [
                "created_at",
                "default",
                "id",
                "movie_dir_template",
                "name",
                "series_season_template",
                "series_show_template",
                "template",
                "type",
                "updated_at"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "default": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "movie_dir_template": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "series_season_template": {
                    "type": "string"
                },
                "series_show_template": {
                    "type": "string"
                },
                "template": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}